# docs/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/games": {
      "get": {
        "description": "Returns global game rows and aggregate tracked-player stats from PostgreSQL only.",
        "operationId": "listGlobalGames",
        "parameters": [
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "name",
                "tracked_players",
                "completion_rate",
                "achievements",
                "playtime"
              ],
              "type": "string"
            }
          },
          {
            "name": "hasAchievements",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalGamesResponseDto"
                }
              }
            }
          }
        },
        "summary": "List tracked Steam games",
        "tags": [
          "games"
        ]
      }
    },
    "/games/{steamAppId}": {
      "get": {
        "description": "Returns canonical game metadata and aggregate tracked-player stats from PostgreSQL only.",
        "operationId": "getGlobalGame",
        "parameters": [
          {
            "name": "steamAppId",
            "required": true,
            "in": "path",
            "schema": {
              "example": 910001,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalGameDetailResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Game has not been tracked."
          }
        },
        "summary": "Get global Steam game detail",
        "tags": [
          "games"
        ]
      }
    },
    "/games/{steamAppId}/achievements": {
      "get": {
        "description": "Returns canonical achievement metadata for one tracked game from PostgreSQL only.",
        "operationId": "listGlobalGameAchievements",
        "parameters": [
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "rarity",
                "name"
              ],
              "type": "string"
            }
          },
          {
            "name": "hidden",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "all",
                "visible",
                "hidden"
              ],
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "steamAppId",
            "required": true,
            "in": "path",
            "schema": {
              "example": 910001,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalGameAchievementsResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Game has not been tracked."
          }
        },
        "summary": "List global game achievement metadata",
        "tags": [
          "games"
        ]
      }
    },
    "/games/{steamAppId}/players": {
      "get": {
        "description": "Returns public Steam profile metadata and progress for tracked players from PostgreSQL only.",
        "operationId": "listGlobalGamePlayers",
        "parameters": [
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "completion",
                "playtime",
                "recently_played"
              ],
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "all",
                "completed",
                "incomplete"
              ],
              "type": "string"
            }
          },
          {
            "name": "steamAppId",
            "required": true,
            "in": "path",
            "schema": {
              "example": 910001,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalGamePlayersResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Game has not been tracked."
          }
        },
        "summary": "List tracked players for a game",
        "tags": [
          "games"
        ]
      }
    },
...
    "/profiles/{steamId}/games/{steamAppId}": {
      "get": {
        "description": "Returns game metadata, profile progress, and summary counts.",
        "operationId": "getProfileGame",
        "parameters": [
          {
            "name": "steamAppId",
            "required": true,
            "in": "path",
            "description": "Steam app ID.",
            "schema": {
              "example": 910002,
              "type": "number"
            }
          },
          {
            "name": "steamId",
            "required": true,
            "in": "path",
            "description": "Steam 64-bit profile ID.",
            "schema": {
              "example": "76561198000000000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDetailResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Profile or profile game row has not been synced."
          }
        },
        "summary": "Get one stored profile game detail",
        "tags": [
          "games"
        ]
      }
    },
    "/profiles/{steamId}/games/{steamAppId}/achievements": {
      "get": {
        "description": "Returns achievement metadata with profile unlock state from PostgreSQL only.",
        "operationId": "listGameAchievements",
        "parameters": [
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "rarity",
                "unlocked_at",
                "name"
              ],
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "all",
                "unlocked",
                "locked"
              ],
              "type": "string"
            }
          },
          {
            "name": "steamAppId",
            "required": true,
            "in": "path",
            "description": "Steam app ID.",
            "schema": {
              "example": 910002,
              "type": "number"
            }
          },
          {
            "name": "steamId",
            "required": true,
            "in": "path",
            "description": "Steam 64-bit profile ID.",
            "schema": {
              "example": "76561198000000000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AchievementsResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Profile or profile game row has not been synced."
          }
        },
        "summary": "List stored achievements for a profile game",
        "tags": [
          "achievements"
        ]
      }
    },
...
    "/leaderboards/{type}": {
      "get": {
        "operationId": "getLeaderboard",
        "parameters": [
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "completion_percentage",
                "completed_games",
                "unlocked_achievements",
                "rarest_unlocks"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Leaderboard type was not found"
          }
        },
        "summary": "Get a leaderboard from latest profile snapshots",
        "tags": [
          "leaderboards"
        ]
      }
    },
...
  },
  "info": {
    "title": "Steam Achievement Tracker API",
    "description": "API for syncing and reading Steam achievement intelligence data",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Docker backend"
    }
  ],
  "components": {
    "schemas": {
...
      "GlobalGameItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "3dd0928a-28f6-4e9b-a5ad-164d536b8d95"
          },
          "steamAppId": {
            "type": "number",
            "example": 910001
          },
          "name": {
            "type": "string",
            "example": "Demo First Steps"
          },
          "iconUrl": {
            "type": "string",
            "example": null,
            "nullable": true
          },
          "logoUrl": {
            "type": "string",
            "example": null,
            "nullable": true
          },
          "hasAchievements": {
            "type": "boolean",
            "example": true
          },
          "trackedPlayers": {
            "type": "number",
            "example": 12
          },
          "totalAchievements": {
            "type": "number",
            "example": 50
          },
          "averageCompletionPercentage": {
            "type": "number",
            "example": 42.34
          },
          "completedPlayers": {
            "type": "number",
            "example": 3
          },
          "totalPlaytimeMinutes": {
            "type": "number",
            "example": 12345
          }
        },
        "required": [
          "id",
          "steamAppId",
          "name",
          "hasAchievements",
          "trackedPlayers",
          "totalAchievements",
          "averageCompletionPercentage",
          "completedPlayers",
          "totalPlaytimeMinutes"
        ]
      },
      "GlobalGamesResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GlobalGameItemResponseDto"
            }
          },
          "total": {
            "type": "number",
            "example": 100
          },
          "limit": {
            "type": "number",
            "example": 25
          },
          "offset": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
      "GlobalGameMetadataResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "3dd0928a-28f6-4e9b-a5ad-164d536b8d95"
          },
          "steamAppId": {
            "type": "number",
            "example": 910001
          },
          "name": {
            "type": "string",
            "example": "Demo First Steps"
          },
          "iconUrl": {
            "type": "string",
            "example": null,
            "nullable": true
          },
          "logoUrl": {
            "type": "string",
            "example": null,
            "nullable": true
          },
          "hasAchievements": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "id",
          "steamAppId",
          "name",
          "hasAchievements"
        ]
      },
      "GlobalGameStatsResponseDto": {
        "type": "object",
        "properties": {
          "trackedPlayers": {
            "type": "number",
            "example": 12
          },
          "completedPlayers": {
            "type": "number",
            "example": 3
          },
          "totalAchievements": {
            "type": "number",
            "example": 50
          },
          "averageCompletionPercentage": {
            "type": "number",
            "example": 42.34
          },
          "totalPlaytimeMinutes": {
            "type": "number",
            "example": 12345
          },
          "averagePlaytimeMinutes": {
            "type": "number",
            "example": 234
          }
        },
        "required": [
          "trackedPlayers",
          "completedPlayers",
          "totalAchievements",
          "averageCompletionPercentage",
          "totalPlaytimeMinutes",
          "averagePlaytimeMinutes"
        ]
      },
      "GlobalGameDetailResponseDto": {
        "type": "object",
        "properties": {
          "game": {
            "$ref": "#/components/schemas/GlobalGameMetadataResponseDto"
          },
          "stats": {
            "$ref": "#/components/schemas/GlobalGameStatsResponseDto"
          }
        },
        "required": [
          "game",
          "stats"
        ]
      },
      "GlobalGameAchievementResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "3dd0928a-28f6-4e9b-a5ad-164d536b8d95"
          },
          "apiName": {
            "type": "string",
            "example": "ACH_WIN_ONE_GAME"
          },
          "displayName": {
            "type": "string",
            "example": "First Win",
            "nullable": true
          },
          "description": {
            "type": "string",
            "example": "Win one match.",
            "nullable": true
          },
          "iconUrl": {
            "type": "string",
            "example": null,
            "nullable": true
          },
          "iconGrayUrl": {
            "type": "string",
            "example": null,
            "nullable": true
          },
          "hidden": {
            "type": "boolean",
            "example": false
          },
          "globalPercentage": {
            "type": "number",
            "example": 12.3,
            "nullable": true
          }
        },
        "required": [
          "id",
          "apiName",
          "hidden"
        ]
      },
      "GlobalGameAchievementsResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GlobalGameAchievementResponseDto"
            }
          },
          "total": {
            "type": "number",
            "example": 50
          },
          "limit": {
            "type": "number",
            "example": 100
          },
          "offset": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
      "GlobalGamePlayerResponseDto": {
        "type": "object",
        "properties": {
          "steamId": {
            "type": "string",
            "example": "76561198000000000"
          },
          "personaName": {
            "type": "string",
            "example": "Player",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "example": null,
            "nullable": true
          },
          "profileUrl": {
            "type": "string",
            "example": null,
            "nullable": true
          },
          "playtimeMinutes": {
            "type": "number",
            "example": 1234
          },
          "totalAchievements": {
            "type": "number",
            "example": 50
          },
          "unlockedAchievements": {
            "type": "number",
            "example": 42
          },
          "completionPercentage": {
            "type": "number",
            "example": 84
          },
          "lastPlayedAt": {
            "type": "string",
            "example": "2026-05-15T12:00:00.000Z",
            "nullable": true
          },
          "publicSlug": {
            "type": "string",
            "example": "nero",
            "nullable": true
          }
        },
        "required": [
          "steamId",
          "playtimeMinutes",
          "totalAchievements",
          "unlockedAchievements",
          "completionPercentage"
        ]
      },
      "GlobalGamePlayersResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GlobalGamePlayerResponseDto"
            }
          },
          "total": {
            "type": "number",
            "example": 12
          },
          "limit": {
            "type": "number",
            "example": 25
          },
          "offset": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
...
      "GameAchievementsSummaryResponseDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 8
          },
          "unlocked": {
            "type": "number",
            "example": 7
          },
          "locked": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "total",
          "unlocked",
          "locked"
        ]
      },
      "GameDetailResponseDto": {
        "type": "object",
        "properties": {
          "steamId": {
            "type": "string",
            "example": "76561198000000000"
          },
          "steamAppId": {
            "type": "number",
            "example": 910002
          },
          "name": {
            "type": "string",
            "example": "Demo Last Mile"
          },
          "iconUrl": {
            "type": "string",
            "example": null,
            "nullable": true
          },
          "logoUrl": {
            "type": "string",
            "example": null,
            "nullable": true
          },
          "hasAchievements": {
            "type": "boolean",
            "example": true
          },
          "playtimeMinutes": {
            "type": "number",
            "example": 1240
          },
          "playtimeTwoWeeksMinutes": {
            "type": "number",
            "example": 45
          },
          "totalAchievements": {
            "type": "number",
            "example": 8
          },
          "unlockedAchievements": {
            "type": "number",
            "example": 7
          },
          "remainingAchievements": {
            "type": "number",
            "example": 1
          },
          "completionPercentage": {
            "type": "number",
            "example": 87.5
          },
          "lastPlayedAt": {
            "type": "string",
            "example": "2026-05-15T12:00:00.000Z",
            "nullable": true
          },
          "lastSyncedAt": {
            "type": "string",
            "example": "2026-05-16T05:07:30.710Z",
            "nullable": true
          },
          "achievementsSummary": {
            "$ref": "#/components/schemas/GameAchievementsSummaryResponseDto"
          }
        },
        "required": [
          "steamId",
          "steamAppId",
          "name",
          "hasAchievements",
          "playtimeMinutes",
          "playtimeTwoWeeksMinutes",
          "totalAchievements",
          "unlockedAchievements",
          "remainingAchievements",
          "completionPercentage",
          "achievementsSummary"
        ]
      },
      "AchievementWithUnlockStateResponseDto": {
        "type": "object",
        "properties": {
          "apiName": {
            "type": "string",
            "example": "ACH_WIN_ONE_GAME"
          },
          "displayName": {
            "type": "string",
            "example": "First Win",
            "nullable": true
          },
          "description": {
            "type": "string",
            "example": "Win one match.",
            "nullable": true
          },
          "iconUrl": {
            "type": "string",
            "example": null,
            "nullable": true
          },
          "iconGrayUrl": {
            "type": "string",
            "example": null,
            "nullable": true
          },
          "globalPercentage": {
            "type": "number",
            "example": 4.321,
            "nullable": true
          },
          "hidden": {
            "type": "boolean",
            "example": false
          },
          "achieved": {
            "type": "boolean",
            "example": true
          },
          "unlockState": {
            "type": "string",
            "enum": [
              "unlocked",
              "locked",
              "unknown"
            ],
            "example": "unlocked",
            "description": "unknown means achievement metadata exists but player unlock state has not been synced or is unavailable."
          },
          "unlockedAt": {
            "type": "string",
            "example": "2026-05-15T12:00:00.000Z",
            "nullable": true
          },
          "lastSyncedAt": {
            "type": "string",
            "example": "2026-05-16T05:07:30.710Z",
            "nullable": true
          }
        },
        "required": [
          "apiName",
          "hidden",
          "achieved",
          "unlockState"
        ]
      },
      "AchievementsResponseDto": {
        "type": "object",
        "properties": {
          "steamId": {
            "type": "string",
            "example": "76561198000000000"
          },
          "steamAppId": {
            "type": "number",
            "example": 910002
          },
          "total": {
            "type": "number",
            "example": 8
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AchievementWithUnlockStateResponseDto"
            }
          }
        },
        "required": [
          "steamId",
          "steamAppId",
          "total",
          "items"
        ]
      },
...
      "LeaderboardSnapshotResponseDto": {
        "type": "object",
        "properties": {
          "totalGames": {
            "type": "number"
          },
          "completedGames": {
            "type": "number"
          },
          "totalAchievements": {
            "type": "number"
          },
          "unlockedAchievements": {
            "type": "number"
          },
          "remainingAchievements": {
            "type": "number"
          },
          "averageCompletionPercentage": {
            "type": "number"
          },
          "rarestUnlockedGlobalPercentage": {
            "type": "number",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "totalGames",
          "completedGames",
          "totalAchievements",
          "unlockedAchievements",
          "remainingAchievements",
          "averageCompletionPercentage",
          "rarestUnlockedGlobalPercentage",
          "createdAt"
        ]
      },
      "LeaderboardItemResponseDto": {
        "type": "object",
        "properties": {
          "rank": {
            "type": "number"
          },
          "steamId": {
            "type": "string"
          },
          "personaName": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "publicSlug": {
            "type": "string",
            "nullable": true
          },
          "score": {
            "type": "number"
          },
          "snapshot": {
            "$ref": "#/components/schemas/LeaderboardSnapshotResponseDto"
          }
        },
        "required": [
          "rank",
          "steamId",
          "personaName",
          "avatarUrl",
          "publicSlug",
          "score",
          "snapshot"
        ]
      },
      "LeaderboardResponseDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "completion_percentage",
              "completed_games",
              "unlocked_achievements",
              "rarest_unlocks"
            ]
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LeaderboardItemResponseDto"
            }
          },
          "limit": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          }
        },
        "required": [
          "type",
          "items",
          "limit",
          "offset"
        ]
      },
...
    }
  }
}